+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 ## Unreleased
 
+### Added
+
+- Add support for configurable formatters with `--format` CLI parameter / `format` config attribute.
+- Add support for JSON formatting with `--format json --quiet`
+
 ## [v0.7.0](https://github.com/thibaudcolas/curlylint/releases/tag/v0.7.0) 2020-04-16
 
 Generally reworked the CLI to match the experience of black.