+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## Unreleased
 
+## [v0.8.0](https://github.com/thibaudcolas/curlylint/releases/tag/v0.8.0) 2020-05-04
+
 ### Added
 
 - Add support for configurable formatters with `--format` CLI parameter / `format` config attribute.